requirements.txt
# aca van los paquetes que necesita el proyecto para andar bien
# si queres instalar todo junto, tirale un pip install -r requirements.txt y listo
# asi te aseguras que no te falte nada y no te vuelvas loco con los errores
Flask>=2.0
flask-httpauth>=4.8
mysql-connector-python>=8.0
pip install python-dotenv
